Update downloadUpdate helper to work with new OOP structure
